requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9 
pydantic-settings==2.1.0 
greenlet==3.0.3 
typing_extensions==4.8.0 
bcrypt==3.2.0 